Codemirror working, basic inspegtor.
It works! You can now edit and run grammars on texts and see the
actions taken by the parser. The AST descent parser is, unsurprisingly,
extremely slow, a fact not helped by doing a DOM manipulation for every
parser action. But hey, if you wanted speed, you should’ve used the
compiled parser. This one is for visibility. :)